Fix router's interface listing view
To get all the ports for a router Horizon uses port_list
filtering by the device_id value which in vanilla openstack
is the quantum router_id, but we use the akanda_id as value for
that field so  the listing doesn't work for us. Lets replace the
port_list call with router_get.

Change-Id: Ibaabb11b2afa97aa2445776842afbdd5c6c2563d
Signed-off-by: Rosario Di Somma <rosario.disomma@dreamhost.com>